# api_specification.yaml
openapi: 3.0.3
info:
  title: Mtaa API
  description: |
    API specification for Mtaa - Hyperlocal Kenyan Neighborhood Social Network
    
    All endpoints require authentication via JWT Bearer token unless otherwise specified.
  version: 1.0.0
  contact:
    name: Mtaa API Support
    email: api@mtaa.app
servers:
  - url: https://api.mtaa.app/api/v1
    description: Production server
  - url: https://staging-api.mtaa.app/api/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        phone_number:
          type: string
        full_name:
          type: string
        username:
          type: string
        profile_image_url:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        neighborhood_id:
          type: string
          format: uuid
        location_verified:
          type: boolean
        verification_status:
          type: string
          enum: [pending, verified, rejected]
        language_preference:
          type: string
          enum: [en, sw]
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        status_code:
          type: integer
...
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Register a new user and send OTP via SMS
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number, full_name, username, neighborhood_id]
              properties:
                phone_number:
                  type: string
                  example: "+254712345678"
                full_name:
                  type: string
                  example: "John Doe"
                username:
                  type: string
                  example: "johndoe"
                neighborhood_id:
                  type: string
                  format: uuid
                address:
                  type: string
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_id:
                    type: string
                    format: uuid
                  message:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: User already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify-otp:
    post:
      tags: [Authentication]
      summary: Verify OTP and get tokens
      description: Verify OTP code and receive JWT tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number, otp_code]
              properties:
                phone_number:
                  type: string
                otp_code:
                  type: string
                  pattern: '^[0-9]{6}$'
      responses:
        '200':
          description: OTP verified successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          description: Invalid OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Request login OTP
      description: Send OTP to user's phone number for login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [phone_number]
              properties:
                phone_number:
                  type: string
      responses:
        '200':
          description: OTP sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /auth/me:
    get:
      tags: [Authentication]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
    
    put:
      tags: [Authentication]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                bio:
                  type: string
                profile_image_url:
                  type: string
                language_preference:
                  type: string
                  enum: [en, sw]
                mpesa_number:
                  type: string
      responses:
        '200':
          description: Profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
...