# apps/backend/src/auth/auth.controller.ts
import { Controller, Post, Body, Get, UseGuards, Request, Delete } from '@nestjs/common';
import { ApiTags, ApiOperation, ApiBearerAuth } from '@nestjs/swagger';
import { AuthService } from './auth.service';
import { RegisterDto, VerifyOtpDto, LoginDto, LoginPasswordDto, RefreshTokenDto } from './dto';
import { JwtAuthGuard } from '../common/guards/jwt-auth.guard';
import { Public } from '../common/decorators/public.decorator';
import { CurrentUser } from '../common/decorators/current-user.decorator';

/**
 * Authentication Controller
 * Handles all authentication-related endpoints
 */
@ApiTags('Authentication')
@Controller('auth')
export class AuthController {
  constructor(private readonly authService: AuthService) {}

  @Public()
  @Post('register')
  @ApiOperation({ summary: 'Register a new user with password' })
  async register(@Body() registerDto: RegisterDto) {
    return this.authService.register(registerDto);
  }

  @Public()
  @Post('register/otp')
  @ApiOperation({ summary: 'Register with OTP (alternative method)' })
  async registerWithOtp(@Body() registerDto: RegisterDto) {
    return this.authService.registerWithOtp(registerDto);
  }

  @Public()
  @Post('login')
  @ApiOperation({ summary: 'Login with password' })
  async login(@Body() loginPasswordDto: LoginPasswordDto) {
    return this.authService.loginWithPassword(loginPasswordDto);
  }

  @Public()
  @Post('login/otp')
  @ApiOperation({ summary: 'Request login OTP' })
  async requestOtp(@Body() loginDto: LoginDto) {
    return this.authService.login(loginDto);
  }

  @Public()
  @Post('verify-otp')
  @ApiOperation({ summary: 'Verify OTP and get tokens' })
  async verifyOtp(@Body() verifyOtpDto: VerifyOtpDto) {
    return this.authService.verifyOtp(verifyOtpDto);
  }

  @Public()
  @Post('refresh-token')
  @ApiOperation({ summary: 'Refresh access token' })
  async refreshToken(@Body() refreshTokenDto: RefreshTokenDto) {
    return this.authService.refreshToken(refreshTokenDto.refreshToken);
  }

  @UseGuards(JwtAuthGuard)
  @Get('me')
  @ApiBearerAuth()
  @ApiOperation({ summary: 'Get current user profile' })
  async getProfile(@CurrentUser() user: any) {
    return this.authService.getProfile(user.userId);
  }

  @UseGuards(JwtAuthGuard)
  @Delete('logout')
  @ApiBearerAuth()
  @ApiOperation({ summary: 'Logout (invalidate refresh token)' })
  async logout(
    @CurrentUser() user: any,
    @Body() body: { refreshToken?: string }
  ) {
    return this.authService.logout(user.userId, body.refreshToken);
  }
}